requirements.txt
aiohttp==3.6.2
aiojobs==0.2.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
idna==2.8
iso8601==0.1.12
kopf==0.22
multidict==4.5.2
pykube-ng==19.10.0
PyYAML==5.1.2
requests==2.22.0
urllib3==1.25.6
yarl==1.3.0
